requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
redis[asyncio]>=5.2.0
openai>=1.54.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
supabase>=2.0.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
